refactor: extract tenant resolution into ravel-tenant-resolve crate
Move the shared tenant-resolution primitives out of ravel-query's HTTP
layer into a new library crate, ravel-tenant-resolve, per ADR-0080
decision 3. This lets ravel-server and the future ravel-ingest-router
depend on one resolver implementation instead of duplicating it, which
is the property canonical-tenant affinity relies on (the two copies
cannot drift).

This is a mechanical, behavior-preserving move. No resolver logic
changes.

The new crate holds TenantResolver, AuthError, OidcError, the concrete
resolvers (DevHeaderTenantResolver, MtlsResolver, OidcJwksCache,
OidcResolver, StaticBearerTokenResolver) and their private helpers, plus
FallbackResolver, which previously lived in ravel-server. The existing
tenant-resolution unit tests move with the code unchanged.

ravel-query keeps its public API surface: the http::tenant module now
re-exports the types from ravel-tenant-resolve, so every existing
ravel_query::http::X path resolves unchanged and no call site is
touched. ravel-server re-exports FallbackResolver from its tenant module
so ravel_server::tenant::FallbackResolver keeps resolving; the
ravel-server-specific orchestration (build_resolver, build_auth_resolver,
ResolverBundle, the JWKS refresh task) stays put.

The now-dead jsonwebtoken and reqwest dependencies of ravel-query, used
only by the moved code, are dropped; blake3 and uuid stay because other
ravel-query modules use them.

Refs: #155
