+++ crates/ravel-query/Cargo.toml
@@ -7,6 +7,12 @@ rust-version.workspace = true
 
 [dependencies]
 ravel-types = { workspace = true }
+# Shared tenant-resolution primitives (ADR-0080 decision 3). ravel-query
+# re-exports these from `http` as its own public API surface
+# (TenantResolver and the concrete resolvers used by AppState); the
+# implementation lives here so ravel-server and ravel-ingest-router share one
+# copy.
+ravel-tenant-resolve = { workspace = true }
 ravel-segment = { workspace = true }
 ravel-logseg = { workspace = true }
 ravel-catalog = { workspace = true }
@@ -35,10 +41,6 @@ tokio = { workspace = true }
 futures = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
-# OIDC tenant resolution (ADR-0042 decision 6): JWT validation and the JWKS
-# fetch behind OidcResolver/OidcJwksCache (crates/ravel-query/src/http/tenant.rs).
-jsonwebtoken = { workspace = true }
-reqwest = { workspace = true }
 # ADR-0071 distributed read fan-out: the queryfrag gRPC surface.
 # `prost` for the message codec, `tonic`/`tonic-prost` for the SeriesFetch
 # service (generated by build.rs). This is a transient inter-process wire
@@ -55,10 +57,9 @@ async-trait = { workspace = true }
 # generates its process id. Already a workspace dependency (previously a
 # dev-dependency here); promoted to a normal dependency, not a new external one.
 uuid = { workspace = true }
-# Constant-secrecy bearer-token lookup (crates/ravel-query/src/http/tenant.rs):
-# configured tokens are stored only as a keyed BLAKE3 hash under a process-local
-# key, mirroring ravel-catalog's auth_token_map, so a raw token is never held or
-# compared in the clear. Already a workspace dependency (used by ravel-catalog);
+# Keyed BLAKE3 MAC over the distributed-read commit token claims
+# (src/distrib/codec.rs) and the content hash of fetched objects
+# (src/distrib/mod.rs). Already a workspace dependency (used by ravel-catalog);
 # promoted to a normal dependency of this crate.
 blake3 = { workspace = true }
 
@@ -90,9 +91,6 @@ criterion = { workspace = true }
 # span test installs a scoped fmt subscriber over a buffer and asserts each
 # phase span fires. Version comes from [workspace.dependencies].
 tracing-subscriber = { workspace = true }
-# The OIDC resolver tests build a local oct (HMAC) JWKS: base64url-encode the
-# signing secret into the JWK `k` field. No network, no keypair generation.
-base64 = { workspace = true }
 
 [[bench]]
 name = "fetch_soa_vs_fetch"
